// web/src/features/assets/assets.types.ts
export type GroupByType = "date" | "type" | "album" | "flat";